JavaDoc modifications only

+++ jipsy-processor/src/main/java/org/kordamp/jipsy/processor/service/ServiceCollector.java
@@ -38,11 +38,8 @@ package org.kordamp.jipsy.processor.service;
 import org.kordamp.jipsy.processor.Initializer;
 import org.kordamp.jipsy.processor.LogLocation;
 import org.kordamp.jipsy.processor.Logger;
-import org.kordamp.jipsy.processor.ProvidedCollector;
 
 import java.util.*;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Maintain a list of services. You can query the services by the name and in case service is not in the list it will
@@ -51,11 +48,11 @@ import static java.util.Objects.requireNonNull;
  * At any time the list of the services can be cached, which is a kind of snapshot and later you can query if there was
  * any modification to the list of the service.
  */
-public final class ServiceCollector implements ProvidedCollector {
-    private final Map<String, Service> services = new LinkedHashMap<>();
-    private final Map<String, Service> cached = new LinkedHashMap<>();
+public final class ServiceCollector {
+    private final Map<String, Service> services = new LinkedHashMap<String, Service>();
+    private final Map<String, Service> cached = new LinkedHashMap<String, Service>();
 
-    private final List<String> removed = new ArrayList<>();
+    private final List<String> removed = new ArrayList<String>();
     private final Initializer initializer;
     private final Logger logger;
 
@@ -76,51 +73,48 @@ public final class ServiceCollector implements ProvidedCollector {
      * @return {@code true} if there was a modification since the service was cached/snapshot made.
      */
     public boolean isModified() {
-        return cached.size() != services.size() ||
-            cached.entrySet().stream().anyMatch(
-                e -> !services.containsKey(e.getKey()) || !e.getValue().equals(services.get(e.getKey())));
-    }
+        if (cached.size() != services.size()) {
+            return true;
+        }
 
-    /**
-     * Get a service for the name of the interface name. If the service is not in the list for this interface name then
-     * load it looking at the accordingly named {@code META-INF/services/xxx} files.
-     *
-     * @param service the fully qualified name of the interface the service providers have to implement.
-     * @return the service instance.
-     */
-    @Override
-    public Service get(String service) {
-        requireNonNull(service,"service");
-        return services.computeIfAbsent(service, this::newService );
-    }
-
-    private Service newService(String name){
-        final Service newService = new Service(logger, name);
-        CharSequence initialData = initializer.initialData(name);
-        if (initialData != null) {
-            newService.fromProviderNamesList(initialData.toString());
-            for (String provider : removed) {
-                newService.removeProvider(provider);
+        for (Map.Entry<String, Service> e : cached.entrySet()) {
+            if (!services.containsKey(e.getKey())) {
+                return true;
+            }
+            if (!e.getValue().equals(services.get(e.getKey()))) {
+                return true;
             }
         }
-        return newService;
+
+        return false;
     }
 
-    /**
-     * @return the collection of the services
-     */
-    @Override
-    public Collection<Service> values() {
+    public Service getService(String service) {
+        if (service == null) {
+            throw new NullPointerException("service");
+        }
+        if (!services.containsKey(service)) {
+            Service newService = new Service(logger, service);
+            CharSequence initialData = initializer.initialData(service);
+            if (initialData != null) {
+                newService.fromProviderNamesList(initialData.toString());
+                for (String provider : removed) {
+                    newService.removeProvider(provider);
+                }
+            }
+            services.put(service, newService);
+        }
+        return services.get(service);
+    }
+
+    public Collection<Service> services() {
         return Collections.unmodifiableMap(services).values();
     }
 
-    /**
-     * Remove a provider from all the services that it provides.
-     *
-     * @param provider the fully qualified name of the class
-     */
     public void removeProvider(String provider) {
-        requireNonNull(provider,"provider");
+        if (provider == null) {
+            throw new NullPointerException("provider");
+        }
         logger.note(LogLocation.LOG_FILE, "Removing " + provider);
         removed.add(provider);
         for (Service service : services.values()) {
